something wrong when receive a QCN

+++ src/traffic-control/model/lossless-queue-disc.cc
@@ -201,7 +201,7 @@ LosslessQueueDisc::DoDequeue (void)
   MyTag tcdTag;
   Ptr<Packet> pk = realitem->GetPacket();
   bool havetag = pk->RemovePacketTag(tcdTag);
-  uint8_t tagval = 0;
+  uint64_t tagval = 0;
   if (havetag){
     tagval |= tcdTag.GetSimpleValue();
   }
